Fix autoloading issue with Rails 4.2 and Sidekiq
Sidekiq was failing to boot with a "Cannot define multiple 'included'
blocks for a Concern" error, caused by the `require` call in the Railtie
(which itself was added so that the server could boot without a missing
constant error).

By wrapping the `on_load` callback in an initializer, we can delay the
callback until the autoload paths have been set up within the app.

See https://stackoverflow.com/a/24486561.